refactor: switch driver to the new implementation
